# go/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GLCharge API documentation.",
        "title": "GLCharge",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/addChargePoint": {
            "post": {
                "parameters": [
                    {
                        "description": "ChargePointReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChargePointReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/addChargePointConnector": {
            "post": {
                "parameters": [
                    {
                        "description": "ChargePointConnectorReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChargePointConnectorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/addGroup": {
            "post": {
                "parameters": [
                    {
                        "description": "GroupReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/changeChargePointPriority": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "ChangeChargePointPriorityReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChangeChargePointPriorityReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HandlerRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/changeConnectorStatus": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "ChangeConnectorStatusReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChangeConnectorStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HandlerRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/changeMaxCurrentGroup": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "ChangeMaxCurrentGroupReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChangeMaxCurrentGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HandlerRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ChangeChargePointPriorityReq": {
            "type": "object",
            "properties": {
                "ChargePointId": {
                    "type": "integer",
                    "example": 1
                },
                "Priority": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handlers.ChangeConnectorStatusReq": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer",
                    "example": 1
                },
                "Status": {
                    "type": "string",
                    "example": "Available"
                }
            }
        },
        "handlers.ChangeMaxCurrentGroupReq": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer",
                    "example": 1
                },
                "MaxCurrent": {
                    "type": "number",
                    "example": 100
                }
            }
        },
        "handlers.ChargePointConnectorReq": {
            "type": "object",
            "properties": {
                "ChargePointId": {
                    "type": "integer",
                    "example": 1
                },
                "Status": {
                    "type": "string",
                    "example": "Available"
                }
            }
        },
        "handlers.ChargePointReq": {
            "type": "object",
            "properties": {
                "GroupId": {
                    "type": "integer",
                    "example": 1
                },
                "Priority": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handlers.GroupReq": {
            "type": "object",
            "properties": {
                "MaxCurrent": {
                    "type": "number",
                    "example": 100
                }
            }
        },
        "handlers.HandlerRes": {
            "type": "object",
            "properties": {
                "resultMap": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                }
            }
        }
    }
}